pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "config-control-plane"
version = "0.1.0"
description = "Centralized config management service with versioning, hot reload, canary rollouts, and audit logs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "jsonschema>=4.23.0,<5.0.0",
  "prometheus-client>=0.22.0,<1.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "redis>=6.0.0,<7.0.0",
  "sqlalchemy>=2.0.40,<3.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "websockets>=15.0.0,<16.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9.0.0",
  "pytest-asyncio>=1.1.0,<2.0.0",
]

[project.scripts]
configctl = "app.cli:main"
config-demo-client = "app.demo_client:main"

[tool.setuptools]
packages = ["app", "app.api", "app.core", "app.db", "app.schemas", "app.services", "app.sdk"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
